testcases/run-parallel.sh: fix parameter passthrough problem

+++ testcases/run-parallel.sh
@@ -35,13 +35,11 @@ done
 
 [[ $# -eq 0 ]] && { Usage; exit 1; }
 distro=$1; shift
-_at=("$@")
-! grep -Eq -- '(^| )(-I=[^ ]+|-[lL])' <<<"$*" && {
-	stdlog=$(trun vm create $distro --downloadonly "$@" |& tee /dev/tty)
-	imgf=$(sed -rn '${/^-[-rwx]{9}.? /{s/^.* //;p}}' <<<"$stdlog")
-	distro=$(awk '/getting fastest location/{print $(NF-1)}' <<<"$stdlog")
-	_at+=($distro "-I=$imgf")
-}
+stdlog=$(trun vm create $distro --downloadonly "$@" |& tee /dev/tty)
+imgf=$(sed -rn '${/^-[-rwx]{9}.? /{s/^.* //;p}}' <<<"$stdlog")
+distro=$(awk '/getting fastest location/{print $(NF-1)}' <<<"$stdlog")
+[[ -z $distro ]] && { echo "{WARN} distro name is empty, exit" >&2; exit; }
+_at=($distro "$@" "-I=$imgf")
 
 for ts in $(tmux ls 2>/dev/null | awk -F: '/fsparallel-test/ {print $1}'); do tmux kill-session -t ${ts}; done
 for ts in $(tmux ls 2>/dev/null | awk -F: '/kissrun-/ {print $1}'); do tmux kill-session -t ${ts}; done
